OnPositionChangedListener를 다시 FloatingTab으로 옮기고, onLayoutChanged 이벤트에 대한 touchView의 이동은 개별 view에서 관리하도록 수정
