+++ hover/src/main/java/io/mattcarroll/hover/FloatingTab.java
@@ -30,6 +30,9 @@ import android.view.ViewGroup;
 import android.view.animation.OvershootInterpolator;
 import android.widget.FrameLayout;
 
+import java.util.Set;
+import java.util.concurrent.CopyOnWriteArraySet;
+
 /**
  * {@code FloatingTab} is the cornerstone of a {@link HoverView}.  When a {@code HoverView} is
  * collapsed, it is reduced to a single {@code FloatingTab} that the user can drag and drop.  When
@@ -42,7 +45,7 @@ import android.widget.FrameLayout;
  *
  * {@code FloatingTab}s position themselves based on their center.
  */
-class FloatingTab extends HoverFrameLayout {
+class FloatingTab extends FrameLayout {
 
     private static final String TAG = "FloatingTab";
     private static final int APPEARING_ANIMATION_DURATION = 300;
@@ -51,11 +54,12 @@ class FloatingTab extends HoverFrameLayout {
     private int mTabSize;
     private View mTabView;
     private Dock mDock;
+    private final Set<OnPositionChangeListener> mOnPositionChangeListeners = new CopyOnWriteArraySet<>();
 
     private final OnLayoutChangeListener mOnLayoutChangeListener = new OnLayoutChangeListener() {
         @Override
         public void onLayoutChange(View v, int left, int top, int right, int bottom, int oldLeft, int oldTop, int oldRight, int oldBottom) {
-            notifyListenersOfPositionChange(FloatingTab.this);
+            notifyListenersOfPositionChange();
         }
     };
 
@@ -225,7 +229,7 @@ class FloatingTab extends HoverFrameLayout {
 
     public void setDock(@NonNull Dock dock) {
         mDock = dock;
-        notifyListenersOfDockChange(mDock);
+        notifyListenersOfDockChange();
     }
 
     public void dock() {
@@ -255,7 +259,7 @@ class FloatingTab extends HoverFrameLayout {
                 if (null != onDocked) {
                     onDocked.run();
                 }
-                notifyListenersOfPositionChange(FloatingTab.this);
+                notifyListenersOfPositionChange();
             }
 
             @Override
@@ -268,7 +272,7 @@ class FloatingTab extends HoverFrameLayout {
         xAnimation.addUpdateListener(new ValueAnimator.AnimatorUpdateListener() {
             @Override
             public void onAnimationUpdate(ValueAnimator animation) {
-                notifyListenersOfPositionChange(FloatingTab.this);
+                notifyListenersOfPositionChange();
             }
         });
     }
@@ -281,7 +285,7 @@ class FloatingTab extends HoverFrameLayout {
         Point cornerPosition = convertCenterToCorner(floatPosition);
         setX(cornerPosition.x);
         setY(cornerPosition.y);
-        notifyListenersOfPositionChange(FloatingTab.this);
+        notifyListenersOfPositionChange();
     }
 
     private Point convertCenterToCorner(@NonNull Point centerPosition) {
@@ -291,9 +295,36 @@ class FloatingTab extends HoverFrameLayout {
         );
     }
 
+    public void addOnPositionChangeListener(@Nullable OnPositionChangeListener listener) {
+        mOnPositionChangeListeners.add(listener);
+    }
+
+    public void removeOnPositionChangeListener(@NonNull OnPositionChangeListener listener) {
+        mOnPositionChangeListeners.remove(listener);
+    }
+
+    private void notifyListenersOfPositionChange() {
+        Point position = getPosition();
+        for (OnPositionChangeListener listener : mOnPositionChangeListeners) {
+            listener.onPositionChange(position);
+        }
+    }
+
+    private void notifyListenersOfDockChange() {
+        for (OnPositionChangeListener listener : mOnPositionChangeListeners) {
+            listener.onDockChange(mDock);
+        }
+    }
+
     // This method is declared in this class simply to make it clear that its part of our public
     // contract and not just an inherited method.
     public void setOnClickListener(@Nullable View.OnClickListener onClickListener) {
         super.setOnClickListener(onClickListener);
+    }
+
+    public interface OnPositionChangeListener {
+        void onPositionChange(@NonNull Point tabPosition);
+
+        void onDockChange(@NonNull Dock dock);
     }
 }
